requirements.txt
dbus-python
PyGObject
psutil
Pillow
